Revert r228650, and work around the clang false positive with printf formats in usr.bin/netstat/atalk.c by conditionally adding NO_WFORMAT to the Makefile instead.
MFC after:	1 week

+++ usr.bin/netstat/atalk.c
@@ -81,7 +81,7 @@ static	char mybuf[50];
 			return("*");
 		}
 	}
-	sprintf(mybuf,"%hu",(short)ntohs(sat->sat_addr.s_net));
+	sprintf(mybuf,"%hu",ntohs(sat->sat_addr.s_net));
 	return mybuf;
 }
